Add support for multiple executeSql statements in transaction that returns results.
If a transaction had more than one "SELECT * FROM TABLE" statement, then only one of the executeSql callbacks would be called.  The others would not be called, even though they were successful.  This is because a transaction object had only one result set.  Changed code so that each executeSql has a result set and unique "query id" so that the query results can be sent back to the correct statement's callback.